Increasing channel history limits
I don't think I'll be deleting submissions, so the channel limit and history limits don't make much sense anymore...

One of the replies uses the channel's history to indicate the submission # and this calculation depends on the history limit which will eventually have to be increased again or removed completely.

The 'Slots' command calculations also depend on the channel history and the limit set.

I will probably try and disable both these features for now.